Add files via upload
Add latest version of code; speed improvements, use Projector, Orthogonality and MUB-constraints to reduce block-sizes. 
Also updated the code which generates blocks for the part of the moment matrix indexed by X_{1,j} (j in [k]) on which S_k acts: this code has been updated and now works in the same way as the full code.